--- teamB/jeongyunah/myapp/db.js
@@ -1,4 +1,0 @@
-export const db_host = process.env.DB_HOST;
-export const db_user = process.env.DB_USER;
-export const db_pass = process.env.DB_PASS;
-
